Files modified from refactoring
