+++ minicluster/src/main/java/org/apache/accumulo/minicluster/cluster/RemoteShellOptions.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.accumulo.cluster;
+package org.apache.accumulo.minicluster.cluster;
 
 import java.io.BufferedReader;
 import java.io.IOException;
