+++ test/src/main/java/org/apache/accumulo/test/ImportExportIT.java
@@ -42,7 +42,7 @@ import java.util.Map.Entry;
 import java.util.Set;
 import java.util.SortedSet;
 
-import org.apache.accumulo.cluster.AccumuloCluster;
+import org.apache.accumulo.minicluster.cluster.AccumuloCluster;
 import org.apache.accumulo.core.Constants;
 import org.apache.accumulo.core.client.Accumulo;
 import org.apache.accumulo.core.client.AccumuloClient;
