+++ test/src/main/java/org/apache/accumulo/test/functional/AccumuloClientIT.java
@@ -27,7 +27,7 @@ import java.util.Map.Entry;
 import java.util.Properties;
 import java.util.Set;
 
-import org.apache.accumulo.cluster.ClusterUser;
+import org.apache.accumulo.minicluster.cluster.ClusterUser;
 import org.apache.accumulo.core.client.Accumulo;
 import org.apache.accumulo.core.client.AccumuloClient;
 import org.apache.accumulo.core.client.BatchWriter;
