+++ test/src/main/java/org/apache/accumulo/test/functional/CloneTestIT_SimpleSuite.java
@@ -39,7 +39,7 @@ import java.util.Set;
 import java.util.TreeSet;
 import java.util.regex.Pattern;
 
-import org.apache.accumulo.cluster.AccumuloCluster;
+import org.apache.accumulo.minicluster.cluster.AccumuloCluster;
 import org.apache.accumulo.core.Constants;
 import org.apache.accumulo.core.client.Accumulo;
 import org.apache.accumulo.core.client.AccumuloClient;
