+++ test/src/main/java/org/apache/accumulo/test/functional/PermissionsIT.java
@@ -35,7 +35,7 @@ import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Set;
 
-import org.apache.accumulo.cluster.ClusterUser;
+import org.apache.accumulo.minicluster.cluster.ClusterUser;
 import org.apache.accumulo.core.client.Accumulo;
 import org.apache.accumulo.core.client.AccumuloClient;
 import org.apache.accumulo.core.client.AccumuloException;
