+++ test/src/main/java/org/apache/accumulo/test/harness/AccumuloClusterHarness.java
@@ -26,11 +26,11 @@ import static org.junit.jupiter.api.Assumptions.assumeTrue;
 import java.io.IOException;
 import java.util.Properties;
 
-import org.apache.accumulo.cluster.AccumuloCluster;
-import org.apache.accumulo.cluster.ClusterControl;
-import org.apache.accumulo.cluster.ClusterUser;
-import org.apache.accumulo.cluster.ClusterUsers;
-import org.apache.accumulo.cluster.standalone.StandaloneAccumuloCluster;
+import org.apache.accumulo.minicluster.cluster.AccumuloCluster;
+import org.apache.accumulo.minicluster.cluster.ClusterControl;
+import org.apache.accumulo.minicluster.cluster.ClusterUser;
+import org.apache.accumulo.minicluster.cluster.ClusterUsers;
+import org.apache.accumulo.minicluster.cluster.standalone.StandaloneAccumuloCluster;
 import org.apache.accumulo.core.client.Accumulo;
 import org.apache.accumulo.core.client.AccumuloClient;
 import org.apache.accumulo.core.client.admin.SecurityOperations;
